backend_requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow==10.0.0
PyMuPDF==1.23.5
python-docx==0.8.11